+++ exporter/exporterhelper/resource_to_label.go
@@ -77,8 +77,8 @@ func addLabelsToMetric(metric *pdata.Metric, labelMap pdata.StringMap) {
 		addLabelsToDoubleDataPoints(metric.DoubleGauge().DataPoints(), labelMap)
 	case pdata.MetricDataTypeIntSum:
 		addLabelsToIntDataPoints(metric.IntSum().DataPoints(), labelMap)
-	case pdata.MetricDataTypeDoubleSum:
-		addLabelsToDoubleDataPoints(metric.DoubleSum().DataPoints(), labelMap)
+	case pdata.MetricDataTypeSum:
+		addLabelsToDoubleDataPoints(metric.Sum().DataPoints(), labelMap)
 	case pdata.MetricDataTypeIntHistogram:
 		addLabelsToIntHistogramDataPoints(metric.IntHistogram().DataPoints(), labelMap)
 	case pdata.MetricDataTypeHistogram: