+++ processor/filterprocessor/expr_test.go
@@ -59,7 +59,7 @@ func TestExprProcessor(t *testing.T) {
 	testFilter(t, pdata.MetricDataTypeIntGauge)
 	testFilter(t, pdata.MetricDataTypeDoubleGauge)
 	testFilter(t, pdata.MetricDataTypeIntSum)
-	testFilter(t, pdata.MetricDataTypeDoubleSum)
+	testFilter(t, pdata.MetricDataTypeSum)
 	testFilter(t, pdata.MetricDataTypeIntHistogram)
 	testFilter(t, pdata.MetricDataTypeHistogram)
 }
@@ -105,8 +105,8 @@ func testFilter(t *testing.T, mdType pdata.MetricDataType) {
 							for l := 0; l < pts.Len(); l++ {
 								assertFiltered(t, pts.At(l).LabelsMap())
 							}
-						case pdata.MetricDataTypeDoubleSum:
-							pts := metric.DoubleSum().DataPoints()
+						case pdata.MetricDataTypeSum:
+							pts := metric.Sum().DataPoints()
 							for l := 0; l < pts.Len(); l++ {
 								assertFiltered(t, pts.At(l).LabelsMap())
 							}